Make user settings responsive & match the styling in other views
