--- src/styles/user/user.css
@@ -1,33 +1,0 @@
-.wrapper {
-  display: flex;
-}
-
-.blueBackground {
-  background: #D9E1E9;
-  border-radius: 24px;
-}
-
-.userSettings {
-  padding: 1rem;
-  width: 720px;
-}
-
-.userFlexRow {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  padding: 1rem 0 0 0;
-}
-
-.userFlexColumn {
-  display: flex;
-  flex-direction: column;
-}
-
-.editButtons {
-  display: flex;
-  align-items: flex-end;
-}
-.editButtons button {
-  margin-left: 0.5rem;
-}
