--- src/styles/user/user.sass
@@ -1,31 +1,0 @@
-$background-blue: #D9E1E9
-$primary-blue:  #0D4D7B
-$light-grey: #6e6c6cb0
-
-.wrapper
-  display: flex
-
-.blueBackground
-  background: $background-blue
-  border-radius: 24px
-
-.userSettings
-  padding: 1rem
-  width: 720px
-
-.userFlexRow
-  display: flex
-  flex-direction: row
-  justify-content: space-between
-  padding: 1rem 0 0 0
-
-.userFlexColumn
-  display: flex
-  flex-direction: column
-
-.editButtons
-  display: flex
-  align-items: flex-end 
-
-  button
-    margin-left: 0.5rem
